+++ lib/karafka/web.rb
@@ -6,6 +6,7 @@
   etc
   open3
   zlib
+  securerandom
 ].each { |lib| require lib }
 
 module Karafka
@@ -31,7 +32,16 @@ module Karafka
       # Activates all the needed routing and sets up listener, etc
       # This needs to run **after** the optional configuration of the web component
       def enable!
+        # Make sure config is as expected
+        # It should be configured before enabling the Web UI
+        Contracts::Config.new.validate!(config.to_h)
+
         Installer.new.enable!
+
+        # Inject correct settings for the Web-UI sessions plugin based on the user configuration
+        # We cannot configure this automatically like other Roda plugins because it requires safe
+        # custom values provided by our user
+        Ui::Base.plugin(:sessions, **config.ui.sessions.to_h)
       end
     end
   end